관리자 - 컨설팅사업관리 기술지도 불량개선이전율, 불량개선이후율 dto 타입 변경, 텍스트 수정

+++ service/src/main/java/com/kap/core/dto/cb/cba/CBATechGuidanceInsertDTO.java
@@ -291,9 +291,9 @@ public class CBATechGuidanceInsertDTO extends BaseDTO {
 
     // 사업진행 상세
     @Schema(title = "불량개선이전율")
-    private Integer fltyImpvmBfreRate;
+    private double fltyImpvmBfreRate;
     @Schema(title = "불량개선이후률")
-    private Integer fltyImpvmAftrRate;
+    private double fltyImpvmAftrRate;
     @Schema(title = "불량개선율")
     private double fltyImpvmRate;
